Started working on taxi coordination

+++ src/main/java/alessio_la_greca_990973/smart_city/SmartCity.java
@@ -64,8 +64,8 @@ public class SmartCity {
         }
     }
 
-    public static int distance(int x1, int y1, int x2, int y2){
-        return (int) Math.ceil(Math.sqrt(Math.pow(x2 - x1, 2) + Math.pow(y2 - y1, 2)));
+    public static double distance(int x1, int y1, int x2, int y2){
+        return Math.ceil(Math.sqrt(Math.pow(x2 - x1, 2) + Math.pow(y2 - y1, 2)));
     }
